+++ src/config/build.rs
@@ -463,6 +463,7 @@ fn build_location(
     let mut access_rules: Vec<(bool, String)> = Vec::new();
     let mut proxy_connect_timeout: Option<std::time::Duration> = None;
     let mut proxy_read_timeout: Option<std::time::Duration> = None;
+    let mut auth_request: Option<Template> = None;
     let mut proxy_cache: Option<String> = None;
     let mut proxy_cache_key: Option<Template> = None;
     let mut proxy_cache_valid: Vec<(Vec<u16>, std::time::Duration)> = Vec::new();
@@ -567,6 +568,11 @@ fn build_location(
                 proxy_read_timeout = Some(parse_duration(&v).ok_or_else(|| {
                     format!("line {}: invalid proxy_read_timeout '{v}'", d.line)
                 })?);
+                None
+            }
+            "auth_request" => {
+                let v = arg1(d)?;
+                auth_request = Some(Template::parse(&v));
                 None
             }
             "include" => None,
@@ -691,6 +697,7 @@ fn build_location(
         proxy_cache_valid,
         proxy_connect_timeout,
         proxy_read_timeout,
+        auth_request,
     })
 }